requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
anthropic>=0.34.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
python-telegram-bot>=20.0
